Complete rewrite to exclusively parse Colloquy XML logs
It turns out that the structured format of the Colloquy logs is easier
to deal with that the X-chat logs, and my users can choose either with
no repurcussions.

This version is equivalent to last night's in terms of infrastructure
and what it can do, but the user experience is much nicer.  In
particular, './ircstats --help' actually does something useful, and it
can output statistics to stdout in CSV format (which the user can then
redirect to anywhere they want.)