Completed: stats persistence, user persistance, user name mapping
Extensive refactoring to increase modularization for the sake or
maintainability.  Stats are now persisted to a pickle file representing
a dictionary of actor objects.  Actor metadata is tracked in a YAML file,
where it can be enhanced by the end-user with, e.g., actor email
addresses, real names, etc.

Added 'validate_yaml', which tries to load the yaml name mapping file
and bombs if it's wrong.  It also documents the file format.