Ignore Templates when generating sumarry statistics
The Wikistats summary reports were producing a lot of lines summarizing
activity stretching back to 2006.  This was confusing, until I realized
that there was activity stretching back that far in the sample data - on
Template pages.  So, now I just don't count any stats for Template
pages.  My extending the Template re with a disjunction, we can also
skip any other content type that we'd like in the future.