More accurate summary stats counts
I looked at the mediawiki source to see what Special:Statistics was
doing to get its counts of content pages.  The filtering its doing is as
follows:
* It skips revisions
* It skips empty pages
* It only gets numbers for things listed as "Content namespaces"

This version skips revisions.  The sample data dumps I have have an
inconsequential number of revisions which are empty, and no pages which
are completely empty.

So any remaining discrepancy is in the "Content namespaces".  I'm
tempted to suggest that something may be misconfigured on the site my
sample stats are drawn from - perhaps the set of content namespaces
isn't very inclusive?

I haven't tried yet, but we could confirm this by counting the number of
entries I find in each namespace and seeing if a subset of namespaces
matches the counts they get.  Perhaps I'll try that after dinner.