Debug output trying to figure out Special:Stats
It's not at all clear from my sample site what Special:Stats on
MediaWiki is actually counting; I'm counting the number of entries in
every valid namespace, and don't come up with numbers anything like what
my sample site has.  At this point, the only thing I can think of is
perhaps the MediaWiki instance, in addition to only counting certain
namespaces, is only counting one entry per title, even when it has a
language designator after it (e.g., /it or /jp).  It's frustrating,
though, and so I'm going to work on something else for a little while.