wikistats: helpful messages to debug only
Wikistats was outputting a few messages that were intended to be helpful
to the end-user, but if the user just wants to capture the output with a
pipe, they annoyingly pollute the output stream.  Rather than calling
these verbose messages, we'll call them debug messages, and send them to
stderr.