TupleDates: First Commit: wikistats speedups
Abandoned instation of datetime objects in favor of passing around
tuples containing portions of the parsed date strings.  They can still
be compared and have math done with them, and their generation should
(hopefully) be much faster.  Also this paves the way to unifying the two
different styles of date cache into one, which should be an improvement
in both time and space complexity.