wikistats: Improved filtering of Proposal namespace
eekim reported that the filter to grab Proposal pages and their related
Talk pages was also grabbing "Proposals", "Proposal How to make wikipedia a
safe encyclopedia", "Proposal for indian market", and also some
Proposals/ pages.  This commit fixes that regular expression so that it
*only* grabs items matching "^Proposal:|^Proposal talk:", making the
namespace explicit and excluding all of those other things.