wikistats: Bug fix on summary statistics
When I switched to TupleDates, I accidentally created a bug in the
summary stats generation, where real date objects were expected.  But
all it did with them was cast them down to strings, so I just built a
string out of the tuples and passed those around instead, and now
everything's happy.