wikistats: fixed reporting for bad dates
When building the list of dates on which events happened, pre-filter
events attached to pages matching a non-content-page regular expression.
Since this date list is used to guide the work of the proposals processor,
this makes its counts much faster and more accurate, in addition to
making the output easier to read.